Moved CreateService to pages
