+++ frontend/dashboard/components/ServiceOwnerSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './ServiceOwnerSelector';
