+++ frontend/dashboard/components/index.ts
@@ -2,3 +2,5 @@ export * from './DataModelsRepoList';
 export * from './FavoriteReposList';
 export * from './OrgRepoList';
 export * from './SearchResultReposList';
+export * from './RepoNameInput';
+export * from './ServiceOwnerSelector';
