+++ frontend/dashboard/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useAugmentReposWithStarred';
-export * from './useRepoService';
