+++ frontend/dashboard/pages/CreateService/index.ts
@@ -1,0 +1,1 @@
+export { CreateService } from './CreateService';
